:recycle: Rename isubscription.ts to subscription.ts
